refactor(movieList): viewholder 패키지 분리
