+++ app/src/main/java/woowacourse/movie/presentation/movieList/adapter/TheaterAdapter.kt
@@ -1,12 +1,11 @@
 package woowacourse.movie.presentation.movieList.adapter
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.model.Theater
+import woowacourse.movie.presentation.movieList.viewholder.TheaterViewHolder
 
 class TheaterAdapter(private val theaters: List<Theater>) :
     RecyclerView.Adapter<TheaterViewHolder>() {
@@ -29,15 +28,5 @@ class TheaterAdapter(private val theaters: List<Theater>) :
         position: Int,
     ) {
         holder.bind(theaters[position])
-    }
-}
-
-class TheaterViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-    private val name: TextView = view.findViewById(R.id.item_theater_name_tv)
-    private val time: TextView = view.findViewById(R.id.item_theater_time_tv)
-
-    fun bind(theater: Theater) {
-        name.text = theater.name
-        time.text = theater.screens.size.toString()
     }
 }
